requirements.txt
fastapi == 0.115.11
uvicorn == 0.34.0
python-dotenv == 1.0.1
mysql-connector-python == 8.0.32
boto3 ==  1.37.11
face_recognition == 1.3.0
numpy == 1.24.3
Pillow == 9.4.0
